+++ laktory/models/dataframe/dataframeexpr.py
@@ -99,13 +99,12 @@ class DataFrameExpr(BaseModel, PipelineChild):
 
     @property
     def data_sources(self):
-        """
-        No inline data sources. All DataFrame references in the SQL expression
-        ({df}, {sources.X}, {nodes.X}) are pre-loaded by PipelineNode.execute()
-        before the transformer runs - {sources.X} from PipelineNode.sources,
-        {nodes.X} via upstream_node_names.
-        """
-        return []
+        """Get PipelineNodeDataSource objects for each {nodes.X} reference in the SQL expression."""
+        from laktory.models.datasources.pipelinenodedatasource import (
+            PipelineNodeDataSource,
+        )
+
+        return [PipelineNodeDataSource(node_name=n) for n in self.upstream_node_names]
 
     def to_sql(self, references=None):
         from laktory.models.datasources.pipelinenodedatasource import (